pyproject.toml
[tool.poetry]
name = "text_classification"
version = "0.1.0"
description = ""
authors = ["Aravindh P"]

[tool.poetry.dependencies]
python = "^3.9"
hydra-core = "^1.1.1"
nltk = "^3.9.1"
spacy = "^3.8.2"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
notebook = "^7.2.2"
huggingface-hub = { extras = ["inference"], version = "^0.25.2" }
pyarrow = "^17.0.0"
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[tool.poetry.dev-dependencies]
pdoc3 = "^0.10.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"
ipykernel = "^6.28.0"
pandas-stubs = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.lint]
ignore = ["E501"]
select = ["B", "C", "E", "F", "W", "B9", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
ignore_missing_imports = true